Skills bug (#7)
* fixed session storage delay

current skill was being retreived from session storage with useEffect i.e. set after component render

fix by initializing current skill with session storage value || 0

broke SkillEntry into separate component for maintainability

* cleanup: consistency

cleaned up workHistory and workEntry formatting to match skills

* renamed SkillsBox to SkillsCarousel

updated imports
