+++ src/app/profile/workhistory/WorkHistory.jsx
@@ -5,7 +5,6 @@ import styles from "./WorkHistory.module.css"
 import { WorkEntry } from "./workEntry";
 import { useEffect, useState } from "react";
 
-
 export const WorkHistory=()=>{
     const stored=sessionStorage.getItem("expanded-entries")
     const [expanded,setExpanded]=useState(stored?JSON.parse(stored):workHistory.map(()=>true))
@@ -14,16 +13,25 @@ export const WorkHistory=()=>{
         prevExpanded.splice(ind,1,bool)
         setExpanded(prevExpanded)
     }
-    useEffect(()=>{
-        storeExpanded(expanded)
-    },[expanded])
-
     const storeExpanded=(exp)=>sessionStorage.setItem("expanded-entries",JSON.stringify(exp))
+    
+    useEffect(() => { storeExpanded(expanded) }, [expanded])
+
     return(
         <div className={styles.container}>
             <h1>Work History</h1>
             <div className={styles.entries}>
-                {workHistory.map((entry,ind)=><WorkEntry key={`${entry.company}_entry`} {...{entry,ind,expanded:expanded[ind],trackExpanded}}/>)}
+                {
+                workHistory.map((entry,ind)=>
+                <WorkEntry key={`${entry.company}_entry`} 
+                {...{
+                    entry,
+                    ind,
+                    expanded:expanded[ind],
+                    trackExpanded
+                }}
+                />)
+                }
             </div>
         </div>
