+++ src/app/profile/workhistory/layout.js
@@ -1,6 +1,5 @@
 import { WorkHistory } from "./WorkHistory";
 
-
 export default function WorkHistoryLayout({children}) {
     return (
       <section>
